F3 Phase B: relocate page dirs under src/pages/
Moved mainview/modelview/render/mocaprender/mocap under src/pages/ (repo root now
holds just src/, webserv/, pdfviewer/, config). pdfviewer stays at root (loaded from
source, self-contained).

Path updates for the now-3-deep page location:
- page HTML classic refs (node_modules/utils/fonts) -> ../../../ (the dist root,
  where they're copied/bundled).
- entry JS imports ../src/X -> ../../X.
- mocap.js (classic) require + MediaPipe locateFile -> ../../../; dropped the dead
  classic vue.js <script> from mocap.html (mocap.js doesn't use Vue).
- holisticPipeline.js locateFile -> ../../../node_modules (page-URL relative; the
  page moved deeper). Both detection paths verified loading MediaPipe.
- vite inputs -> src/pages/*; main.js loadFile + the discrete BrowserView path ->
  dist/src/pages/*.
- models/ + pdfs/ copied to dist/src/pages/ (beside the pages) so the ../models and
  ../pdfs runtime-string refs resolve; models.json kept "../models/..." so the web
  server's path.resolve(__dirname, modelObj.path) still finds repo/models.

dist now mirrors the source (dist/src/pages/<page>/). Verified: model library +
thumbnails, model viewer, settings, mocap (integrated + discrete), MediaPipe.
