+++ main.js
@@ -109,7 +109,7 @@ function createWindow() {
     // ../mocaprender/...) and the discrete BrowserView still resolve, and all
     // pages share one file:// origin for the localStorage handoff (Invariant 1).
     // Run `npm run build` first (npm start does this). See docs/REFACTOR.md.
-    mainWindow.loadFile("dist/mainview/framework.html");
+    mainWindow.loadFile("dist/src/pages/mainview/framework.html");
     if(storage.getItem("useDark")){
         nativeTheme.themeSource = "dark";
     }else{
@@ -317,7 +317,7 @@ function createModelViewerWindow(args) {
 
     viewer.webContents.once('dom-ready', () => {
         viewer.show();
-        viewer.loadFile("dist/modelview/modelview.html");
+        viewer.loadFile("dist/src/pages/modelview/modelview.html");
     });
 
     // Open the DevTools.
