+++ src/pages/mainview/framework.html
@@ -17,15 +17,15 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <title>SysMocap</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="../fonts/fonts.css" />
+    <link rel="stylesheet" href="../../../fonts/fonts.css" />
     <link rel="stylesheet" href="style.css" />
     <!-- import thirdpart library -->
-    <link rel="stylesheet" href="../node_modules/mdui/dist/css/mdui.css" />
-    <script src="../node_modules/mdui/dist/js/mdui.js"></script>
+    <link rel="stylesheet" href="../../../node_modules/mdui/dist/css/mdui.css" />
+    <script src="../../../node_modules/mdui/dist/js/mdui.js"></script>
     <!-- Vue 3 is ESM-bundled via framework.js -->
     <!-- window globals used imperatively by App.vue's domBoom: html2canvas + ParticleBoom -->
-    <script src="../utils/html2canvas.js"></script>
-    <script src="../utils/particle-boom.umd.js"></script>
+    <script src="../../../utils/html2canvas.js"></script>
+    <script src="../../../utils/particle-boom.umd.js"></script>
 </head>
 
 <body style="font-family: 'Quicksand', 'NowarRounded'">
