+++ src/pages/mocap/mocap.html
@@ -22,12 +22,11 @@
         <!-- Legacy three.js / three-vrm global builds removed: mocap.js is
              capture-only (MediaPipe + Kalidokit) and never used them. -->
         <!-- Mediapipe -->
-        <script src="../node_modules/@mediapipe/holistic/holistic.js"></script>
+        <script src="../../../node_modules/@mediapipe/holistic/holistic.js"></script>
         <!-- Mediapipe Drawing Tools -->
-        <script src="../node_modules/@mediapipe/drawing_utils/drawing_utils.js"></script>
+        <script src="../../../node_modules/@mediapipe/drawing_utils/drawing_utils.js"></script>
         <!-- Mediapipe Camera Tools -->
-        <script src="../node_modules/@mediapipe/camera_utils/camera_utils.js"></script>
-        <script src="../node_modules/vue/dist/vue.js"></script>
+        <script src="../../../node_modules/@mediapipe/camera_utils/camera_utils.js"></script>
         <style>
             ::-webkit-scrollbar {
                 display: none;
@@ -72,7 +71,7 @@
             ></video>
         </div>
 
-        <script src="../node_modules/kalidokit/dist/kalidokit.umd.js"></script>
+        <script src="../../../node_modules/kalidokit/dist/kalidokit.umd.js"></script>
         <script src="./mocap.js" defer></script>
     </body>
 </html>
