+++ src/pages/mocaprender/render.html
@@ -21,12 +21,12 @@
         />
 
         <!-- MediaPipe (detection runs in this page; holisticPipeline uses these globals) -->
-        <script src="../node_modules/@mediapipe/holistic/holistic.js"></script>
-        <script src="../node_modules/@mediapipe/drawing_utils/drawing_utils.js"></script>
-        <script src="../node_modules/@mediapipe/camera_utils/camera_utils.js"></script>
+        <script src="../../../node_modules/@mediapipe/holistic/holistic.js"></script>
+        <script src="../../../node_modules/@mediapipe/drawing_utils/drawing_utils.js"></script>
+        <script src="../../../node_modules/@mediapipe/camera_utils/camera_utils.js"></script>
         <link rel="stylesheet" href="style.css" />
-        <link rel="stylesheet" href="../node_modules/mdui/dist/css/mdui.css" />
-        <script src="../node_modules/mdui/dist/js/mdui.js"></script>
+        <link rel="stylesheet" href="../../../node_modules/mdui/dist/css/mdui.css" />
+        <script src="../../../node_modules/mdui/dist/js/mdui.js"></script>
     </head>
 
     <body class="mdui-theme-layout-auto" style="background-color: #0000">
@@ -34,9 +34,9 @@
 
         <!-- window globals used imperatively: html2canvas + RecordRTC (recording),
              Kalidokit (solve + helpers). three / @pixiv/three-vrm are ESM-bundled. -->
-        <script src="../utils/html2canvas.js"></script>
-        <script src="../utils/RecordRTC.js"></script>
-        <script src="../node_modules/kalidokit/dist/kalidokit.umd.js"></script>
+        <script src="../../../utils/html2canvas.js"></script>
+        <script src="../../../utils/RecordRTC.js"></script>
+        <script src="../../../node_modules/kalidokit/dist/kalidokit.umd.js"></script>
         <script src="./script.js" defer type="module"></script>
     </body>
 </html>
