+++ src/pages/modelview/modelview.html
@@ -8,13 +8,13 @@
             content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no"
         />
         <title>ModelVierer</title>
-        <link rel="stylesheet" href="../node_modules/mdui/dist/css/mdui.css" />
-        <script src="../node_modules/mdui/dist/js/mdui.js"></script>
+        <link rel="stylesheet" href="../../../node_modules/mdui/dist/css/mdui.css" />
+        <script src="../../../node_modules/mdui/dist/js/mdui.js"></script>
         <!-- Vue 3 + three/vrm/svelte-jsoneditor are ESM-bundled via modelview.js -->
 
-        <script src="../node_modules/lil-gui/dist/lil-gui.umd.min.js"></script>
+        <script src="../../../node_modules/lil-gui/dist/lil-gui.umd.min.js"></script>
 
-        <link rel="stylesheet" href="../fonts/fonts.css" />
+        <link rel="stylesheet" href="../../../fonts/fonts.css" />
 
         <style>
             body {
