+++ src/pages/render/render.html
@@ -21,8 +21,8 @@
         />
 
         <link rel="stylesheet" href="render.css" />
-        <link rel="stylesheet" href="../node_modules/mdui/dist/css/mdui.css" />
-        <script src="../node_modules/mdui/dist/js/mdui.js"></script>
+        <link rel="stylesheet" href="../../../node_modules/mdui/dist/css/mdui.css" />
+        <script src="../../../node_modules/mdui/dist/js/mdui.js"></script>
     </head>
 
     <body class="mdui-theme-layout-auto" style="background-color: #0000">
@@ -31,9 +31,9 @@
         <!-- window globals used imperatively by avatarRenderer:
              html2canvas + RecordRTC (recording), Kalidokit (Utils/Vector helpers).
              three / @pixiv/three-vrm are ESM-bundled via avatarRenderer's imports. -->
-        <script src="../utils/html2canvas.js"></script>
-        <script src="../utils/RecordRTC.js"></script>
-        <script src="../node_modules/kalidokit/dist/kalidokit.umd.js"></script>
+        <script src="../../../utils/html2canvas.js"></script>
+        <script src="../../../utils/RecordRTC.js"></script>
+        <script src="../../../node_modules/kalidokit/dist/kalidokit.umd.js"></script>
         <script src="./render.js" defer type="module"></script>
     </body>
 </html>
